requirements.txt
hydra-core==1.3.2
numpy==1.26.4
pandas==2.1.4
scipy==1.11.4
torch==2.2.1
matplotlib==3.8.2
plotly==5.18.0
pyyaml==6.0.1
wandb==0.16.2
opt_einsum==3.3.0

ruff==0.1.9
